feat: add logic for traceParentInHeader + tests
